refactor : InputView역할로 리펙토링 MVC모델에서 VIEW부분에 있는 여러 Input파일을 하나의 InputView 객체로 리펙토링, 그리고 유효성 검사 부분 분리
